:pencil2: [Fix] userId를 Long 타입으로 변경, memberId -> userId로 변경

+++ src/main/java/com/ovtrip/global/jwt/dto/RefreshTokenDto.java
@@ -7,12 +7,12 @@ import java.time.LocalDateTime;
 
 @Data
 public class RefreshTokenDto {
-    private int userId;
+    private Long userId;
     private String refreshToken;
     private LocalDateTime refreshTokenExpireTime;
 
     @Builder
-    public RefreshTokenDto(int userId, String refreshToken, LocalDateTime refreshTokenExpireTime) {
+    public RefreshTokenDto(Long userId, String refreshToken, LocalDateTime refreshTokenExpireTime) {
         this.userId = userId;
         this.refreshToken = refreshToken;
         this.refreshTokenExpireTime = refreshTokenExpireTime;
