+++ src/main/java/com/ovtrip/user/model/mapper/UserMapper.java
@@ -1,6 +1,5 @@
 package com.ovtrip.user.model.mapper;
 
-import com.ovtrip.global.jwt.dto.JwtTokenDto;
 import com.ovtrip.global.jwt.dto.RefreshTokenDto;
 import com.ovtrip.user.model.dto.LoginUserDto;
 import com.ovtrip.user.model.dto.UserDto;
@@ -16,9 +15,9 @@ public interface UserMapper {
     // 회원가입시 이메일 중복 체크
     int idCheck(String userEmail) throws SQLException;
     // 회원가입
-    int joinUser(UserDto userDto) throws SQLException;
+    Long joinUser(UserDto userDto) throws SQLException;
     // 회원 아이디로 회원 조회
-    UserVO getUserById(int userId) throws SQLException;
+    UserVO getUserById(Long userId) throws SQLException;
     // 회원 이메일로 회원 조회
     UserVO getUserByEmail(String userEmail) throws SQLException;
     // 로그인
